Provide repository to handle WP Toolkit
Handle the upgrade of the third-party WP Toolkit software when it is
detected. Currently, this only works for installations done through
the cPanel control panel, but the framework for providing support can be
extended to a standalone installation if the WP Toolkit software comes
to support this in the future.

The support will not be activated unless the accompanying `wp-toolkit`
vendor, defined in the `leapp-data` repository, is available and active.