Parameterize repo used during upgrade
The reason for this is that the so-called "latest" repo for WP Toolkit
doesn't always instruct yum/dnf to install the actual latest software.
If it does this, WPTK will be downgraded instead. To avoid this,
populate a variable under /etc/dnf/vars/ and use that variable in the
repo used by the upgrade.

Also, refactor a lot to make this reasonable to do.