Relax check_panel_memory actor's check
The check_panel_memory actor is intended to inhibit upgrades when the
system lacks the memory needed to satisfy minimum system requirements
for CloudLinux 8, which may vary depending on whether a control panel is
installed. However, the actor is using the value for total memory as
reported by the kernel, rather than the nominal amount of memory
actually installed on the system. Since the kernel subtracts its own
memory usage from this total, a nominally minimum-spec machine will
always fail this check. To address this, credit the machine with a
certain amount of additional memory when performing the check. (This
credited amount will not appear in the message if the check still
fails.)