Add preliminary repository for Microsoft software
Microsoft publishes an RPM in their repos which provides a canonical YUM
repo file; however, it seems to be common practice not to use this
package-owned file. This makes the upgrade more trouble than it should
be. Instead, make sure that the package is installed during the update,
and then rename the old manual repo file to disable it on first boot
into the new OS.

The actors all depend on the accompanying PES vendor in order to do
anything.